+++ src/Falcor/NumericKey.cs
@@ -3,7 +3,7 @@
     public abstract class NumericKey : KeySegment
     {
 
-        public static implicit operator NumericKey(long value)
+        public static implicit operator NumericKey(int value)
         {
             return new NumberKey(value);
         }
